cdk/requirements.txt
aws-cdk-lib==2.80.0
constructs>=10.0.0
boto3>=1.26.0
pytest>=7.0.0           # opcional para tests
pip-tools               # opcional, para gestionar dependencias
